# Optimised trie to use an array instead of a map. Added benchmarks.