--- a/pkg/rules/trie.go
+++ b/pkg/rules/trie.go
@@ -3,15 +3,13 @@ package rules
 // PrefixTrie efficiently checks if a slice of bytes contains one of the prefixes prevoiusly
 // added to the tree.
 type PrefixTrie struct {
-	nxt map[byte]*PrefixTrie
+	nxt [256]*PrefixTrie
 	end bool
 }
 
 // NewPrefixTrie makes a new instance of a PrefixTrie.
 func NewPrefixTrie() *PrefixTrie {
-	return &PrefixTrie{
-		nxt: make(map[byte]*PrefixTrie),
-	}
+	return &PrefixTrie{}
 }
 
 // Add a prefix to the trie
@@ -19,9 +17,7 @@ func (t *PrefixTrie) Add(s []byte) {
 	n := t
 	for _, c := range s {
 		if n.nxt[c] == nil {
-			n.nxt[c] = &PrefixTrie{
-				nxt: make(map[byte]*PrefixTrie),
-			}
+			n.nxt[c] = &PrefixTrie{}
 		}
 		n = n.nxt[c]
 	}