Changed trie to use an array instead of a map.

The benchmark in 03a3b8aad027ff314f605e9cfa0f83fe027e2e6b gives

```
name    old time/op  new time/op  delta
Trie-8  6.55ms ± 2%  0.29ms ± 5%  -95.64%  (p=0.000 n=10+9)
```